[BUGFIX] rename reviewAdminItem, reviewMemberItem, reviewPublicItem, refactor CommonModal type
