+++ src/components/review/ReviewMemberItemCollection.tsx
@@ -1,22 +1,42 @@
 import { useQuery } from '@apollo/react-hooks'
 import { Box, Button, SkeletonCircle, SkeletonText } from '@chakra-ui/react'
 import gql from 'graphql-tag'
-import React, { useState } from 'react'
+import React, { HTMLAttributes, useState } from 'react'
 import { useIntl } from 'react-intl'
 import { commonMessages } from '../../helpers/translation'
 import types from '../../types'
 import { ReviewLabelRoleProps, ReviewProps } from '../../types/review'
+import { useAuth } from '../auth/AuthContext'
 import { StyledDivider } from './ReviewCollectionBlock'
 import ReviewItem from './ReviewItem'
 
-const ReviewPublicItem: React.FC<{
-  targetId: string
-  path: string
-  appId: string
-}> = ({ path, appId, targetId }) => {
+export interface ReviewMemberItemRef {
+  onRefetchReviewMemberItem: () => void
+}
+
+const ReviewMemberItemCollection: React.ForwardRefRenderFunction<
+  ReviewMemberItemRef,
+  HTMLAttributes<HTMLDivElement> & {
+    targetId: string
+    path: string
+    appId: string
+  }
+> = ({ path, appId, targetId }, ref) => {
   const { formatMessage } = useIntl()
+  const { currentMemberId } = useAuth()
   const [loading, setLoading] = useState(false)
-  const { loadingReviews, publicReviews, labelRole, loadMoreReviews } = useReviewPublicCollection(path, appId, targetId)
+  const {
+    loadingReviews,
+    memberReviews,
+    memberPrivateContent,
+    labelRole,
+    onRefetch,
+    loadMoreReviews,
+  } = useReviewMemberCollection(path, appId, currentMemberId, targetId)
+
+  React.useImperativeHandle(ref, () => ({
+    onRefetchReviewMemberItem: () => onRefetch(),
+  }))
 
   if (loadingReviews) {
     return (
@@ -29,7 +49,7 @@ const ReviewPublicItem: React.FC<{
 
   return (
     <>
-      {publicReviews.map(v => (
+      {memberReviews.map(v => (
         <div key={v.id} className="review-item">
           <ReviewItem
             id={v.id}
@@ -41,6 +61,13 @@ const ReviewPublicItem: React.FC<{
             updatedAt={v.updatedAt}
             reviewReplies={v.reviewReplies}
             labelRole={labelRole}
+            privateContent={
+              memberPrivateContent &&
+              memberPrivateContent.length !== 0 &&
+              v.memberId === memberPrivateContent[0].memberId
+                ? memberPrivateContent[0].privateContent
+                : null
+            }
           />
           <StyledDivider className="review-divider" />
         </div>
@@ -63,17 +90,24 @@ const ReviewPublicItem: React.FC<{
   )
 }
 
-const useReviewPublicCollection = (path: string, appId: string, targetId: string) => {
-  const condition: types.GET_REVIEW_PUBLICVariables['condition'] = {
+const useReviewMemberCollection = (path: string, appId: string, currentMemberId: string | null, targetId: string) => {
+  const condition: types.GET_REVIEW_MEMBERVariables['condition'] = {
     path: { _eq: path },
     app_id: { _eq: appId },
   }
+
   const { loading, error, data, refetch, fetchMore } = useQuery<
-    types.GET_REVIEW_PUBLIC,
-    types.GET_REVIEW_PUBLICVariables
+    types.GET_REVIEW_MEMBER,
+    types.GET_REVIEW_MEMBERVariables
   >(
     gql`
-      query GET_REVIEW_PUBLIC($condition: review_public_bool_exp, $targetId: uuid, $limit: Int!) {
+      query GET_REVIEW_MEMBER(
+        $condition: review_public_bool_exp
+        $path: String!
+        $currentMemberId: String
+        $targetId: uuid
+        $limit: Int!
+      ) {
         review_public_aggregate(where: $condition) {
           aggregate {
             count
@@ -98,6 +132,11 @@ const useReviewPublicCollection = (path: string, appId: string, targetId: string
             }
           }
         }
+        review(where: { path: { _eq: $path }, member_id: { _eq: $currentMemberId } }) {
+          id
+          member_id
+          private_content
+        }
         program_role(where: { program_id: { _eq: $targetId } }) {
           id
           name
@@ -108,13 +147,15 @@ const useReviewPublicCollection = (path: string, appId: string, targetId: string
     {
       variables: {
         condition,
+        path,
+        currentMemberId,
         targetId,
         limit: 5,
       },
     },
   )
 
-  const publicReviews: ReviewProps[] =
+  const memberReviews: ReviewProps[] =
     data?.review_public.map(v => ({
       id: v.id,
       memberId: v.member_id,
@@ -132,6 +173,14 @@ const useReviewPublicCollection = (path: string, appId: string, targetId: string
         updatedAt: new Date(v.updated_at),
       })),
     })) || []
+
+  const memberPrivateContent =
+    loading || error || !data
+      ? null
+      : data.review.map(v => ({
+          memberId: v.member_id,
+          privateContent: v.private_content,
+        }))
 
   const labelRole: ReviewLabelRoleProps[] =
     loading || error || !data
@@ -162,12 +211,13 @@ const useReviewPublicCollection = (path: string, appId: string, targetId: string
 
   return {
     loadingReviews: loading,
-    errorReviews: error,
-    publicReviews,
+    error,
+    memberReviews,
+    memberPrivateContent,
     labelRole,
     onRefetch: refetch,
     loadMoreReviews: (data?.review_public_aggregate.aggregate?.count || 0) > 5 ? loadMoreReviews : undefined,
   }
 }
 
-export default ReviewPublicItem
+export default React.forwardRef(ReviewMemberItemCollection)
